Add Balance

+++ app/src/main/java/com/succiue/myapplication/data/model/AccountModel.kt
@@ -1,10 +1,10 @@
 package com.succiue.myapplication.data.model
 
 data class BalanceModel(
-    val available: Long?,
-    val current: Long?,
+    val available: Double?,
+    val current: Double?,
     val iso_currency_code: String?,
-    val limit: Long?,
+    val limit: Double?,
     val unofficial_currency_code: String?
 )
